refactor: split FileFilter out from build.js to util/file-filter.js
