+++ tests/unit/test-cmd/test.build.js
@@ -8,9 +8,9 @@ import sinon from 'sinon';
 
 import build, {
   safeFileName,
-  FileFilter,
   getDefaultLocalizedName,
 } from '../../../src/cmd/build';
+import {FileFilter} from '../../../src/util/file-filter';
 import {withTempDir} from '../../../src/util/temp-dir';
 import {
   fixturePath,
